test(python): Use pytest marker for slow tests
